:wrench: :boom: Overhaul build tools and processes
Switch from Sass to postCSS for building plugin styles and add
build tools for working with the WP block editor -- namely Webpack
and Babel. Also upgrade to a newer phpcs setup (still using
Composer for package management).

This includes new and updated configuration files for a variety
of test and build tools.